Values of the game settings are read from the grid
Previously, the values for the height and width of the field and the number of mines were specified as fixed values in the input fields. This caused the values to be reset after restarting the game. Due to the change in the code, the actual values are now always displayed.